fix: amend licensing
This pull request updates the licensing terms of the project from the MIT License to a dual-license model under both the MIT License and the Apache License 2.0.
This change aims to provide greater flexibility and compatibility for users and contributors in the rust ecosystem, where dual-licensing under MIT/Apache2 is common.

The following prior authors of this library need to agree.

- @h4llow3En
- @hoodie
- @pandawan
- @BlackHoleFox
- @betamos
- @mihirsamdarshi
- @amrbashir
- @frewsxcv
- @Guiguiprim
- @JoshMcguigan
- @probablykasper
- @lucasfernog
- @nazar-pc
- @andytom
- @mkroehnert

Contributor checkoff
--------------------

To agree to relicensing, comment with :

I license past and future contributions under the dual MIT/Apache-2.0 license, allowing licensees to chose either at their option

Or, if you're a contributor, you can check the box in this repo next to your
name. My scripts will pick this exact phrase up and check your checkbox, but
I'll come through and manually review this issue later as well.